runtime: Move compile-conditins utility ot
