+++ packages/runtime/src/execute/plan.ts
@@ -9,8 +9,7 @@ import type {
 } from '../types';
 import clone from '../util/clone';
 import validatePlan from '../util/validate-plan';
-import compileFunction from '../modules/compile-function';
-import { preconditionContext } from './context';
+import compileConditions from './compile-conditions';
 
 type ExeContext = {
   plan: CompiledExecutionPlan;
@@ -31,42 +30,6 @@ const assembleState = (
   data: Object.assign({}, initialState.data || {}, data),
 });
 
-// TODO move out of here
-export const compileConditions = (plan: ExecutionPlan) => {
-  const context = preconditionContext();
-
-  if (plan.precondition && typeof plan.precondition === 'string') {
-    try {
-      (plan as CompiledExecutionPlan).precondition = compileFunction(
-        plan.precondition,
-        context
-      );
-    } catch (e: any) {
-      throw new Error(`Failed to compile plan precondition (${e.message})`);
-    }
-  }
-  for (const jobId in plan.jobs) {
-    const job = plan.jobs[jobId];
-    if (job.next) {
-      for (const edgeId in job.next) {
-        try {
-          const edge = job.next[edgeId];
-          if (edge.condition && typeof edge.condition === 'string') {
-            edge.condition = compileFunction(edge.condition, context);
-          }
-        } catch (e: any) {
-          throw new Error(
-            `Failed to compile edge condition on ${jobId}-${edgeId}(${e.message})`
-          );
-        }
-      }
-    }
-  }
-  return plan as CompiledExecutionPlan;
-};
-
-// TODO accept initial state
-// On the first job, merge state and initialState
 const executePlan = async (
   plan: ExecutionPlan,
   initialState: State = {},
